+++ cmds/start-testnet-voting.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 set -e
-symbol-bootstrap start -p testnet -a dual -t target/testnet -c test/voting_preset.yml  $1
+symbol-bootstrap start -p testnet -a dual -t target/testnet -c test/voting_preset.yml  $1 $2 $3